+++ ceammc/CHANGELOG.md
@@ -19,6 +19,7 @@
 - ui.hgain~ alias added to create horizontal ui.gain~
 - ui.hgain2~ alias added to create horizontal ui.gain2~
 - ui.polar added
+- popup menu added to ui.preset with read/write actions
 
 ### Changed:
 - ui.tab in single mode outputs like ui.menu: list pair - INDEX TAB_NAME